+++ src/native/data/DataName.ts
@@ -46,4 +46,6 @@ export class DataName {
     
     static EVENT_NAME = "eventName";
     static EVENT_VALUES = "eventValues";
+    static PUSH_TOKEN = "push_token";
+    static PUSH_TOKEN_SERVICE = "push_token_service";
 }