+++ src/settings/PushTokenService.ts
@@ -1,0 +1,4 @@
+export enum PushTokenService { 
+    FIREBASE = "fms",
+    APPLE = "apns",
+}